refact: 함수 모듈화
